Getting ready for demo

+++ CosmosDB/Models/Order.cs
@@ -8,7 +8,7 @@ public class Order
     public DateTime? RequiredDate { get; set; }
     public DateTime? ShippedDate { get; set; }
     public int? ShipVia { get; set; }
-    public decimal? Freight { get; set; }
+    public double? Freight { get; set; }
     public string? ShipName { get; set; }
     public string? ShipAddress { get; set; }
     public string? ShipCity { get; set; }
